requirements.txt
PySide6>=6.6.0
opencv-python>=4.9.0
pyautogui>=0.9.54
Pillow>=10.0.0
numpy>=1.24.0
